@@ -72,19 +72,27 @@ func (a MempoolTxidEntries) Less(i, j int) bool {
 // removeEntryFromMempool removes entry from mempool structs. The caller is responsible for locking!
 func (m *BaseMempool) removeEntryFromMempool(txid string, entry txEntry) {
 	delete(m.txEntries, txid)
+	// store already processed addrDesc - it can appear multiple times as a different outpoint
+	processedAddrDesc := make(map[string]struct{})
 	for _, si := range entry.addrIndexes {
 		outpoints, found := m.addrDescToTx[si.addrDesc]
 		if found {
-			newOutpoints := make([]Outpoint, 0, len(outpoints)-1)
-			for _, o := range outpoints {
-				if o.Txid != txid {
-					newOutpoints = append(newOutpoints, o)
+			_, processed := processedAddrDesc[si.addrDesc]
+			if !processed {
+				processedAddrDesc[si.addrDesc] = struct{}{}
+				j := 0
+				for i := 0; i < len(outpoints); i++ {
+					if outpoints[i].Txid != txid {
+						outpoints[j] = outpoints[i]
+						j++
+					}
 				}
-			}
-			if len(newOutpoints) > 0 {
-				m.addrDescToTx[si.addrDesc] = newOutpoints
-			} else {
-				delete(m.addrDescToTx, si.addrDesc)
+				outpoints = outpoints[:j]
+				if len(outpoints) > 0 {
+					m.addrDescToTx[si.addrDesc] = outpoints
+				} else {
+					delete(m.addrDescToTx, si.addrDesc)
+				}
 			}
 		}
 	}
