@@ -7,12 +7,9 @@ import (
 	"net/url"
 
 	jsontypes "github.com/ava-labs/avalanchego/utils/json"
-	"github.com/ava-labs/coreth/core/types"
-	"github.com/ava-labs/coreth/ethclient"
-	"github.com/ava-labs/coreth/interfaces"
-	"github.com/ava-labs/coreth/rpc"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/golang/glog"
 	"github.com/juju/errors"
 	"github.com/trezor/blockbook/bchain"
@@ -52,7 +49,7 @@ func (b *AvalancheRPC) Initialize() error {
 			return nil, nil, err
 		}
 		rc := &AvalancheRPCClient{Client: r}
-		c := &AvalancheClient{Client: ethclient.NewClient(r)}
+		c := &AvalancheClient{Client: ethclient.NewClient(r), AvalancheRPCClient: rc}
 		return rc, c, nil
 	}
 
@@ -81,7 +78,7 @@ func (b *AvalancheRPC) Initialize() error {
 	b.RPC = rpcClient
 	b.info = infoClient
 	b.MainNetChainID = MainNet
-	b.NewBlock = &AvalancheNewBlock{channel: make(chan *types.Header)}
+	b.NewBlock = &AvalancheNewBlock{channel: make(chan *Header)}
 	b.NewTx = &AvalancheNewTx{channel: make(chan common.Hash)}
 
 	ctx, cancel := context.WithTimeout(context.Background(), b.Timeout)
@@ -101,6 +98,8 @@ func (b *AvalancheRPC) Initialize() error {
 		return errors.Errorf("Unknown network id %v", id)
 	}
 
+	b.InitAlternativeProviders()
+
 	glog.Info("rpc: block chain ", b.Network)
 
 	return nil
@@ -110,7 +109,6 @@ func (b *AvalancheRPC) Initialize() error {
 func (b *AvalancheRPC) GetChainInfo() (*bchain.ChainInfo, error) {
 	ci, err := b.EthereumRPC.GetChainInfo()
 	if err != nil {
-		fmt.Println(err)
 		return nil, err
 	}
 
@@ -125,40 +123,11 @@ func (b *AvalancheRPC) GetChainInfo() (*bchain.ChainInfo, error) {
 		VMVersions         map[string]string `json:"vmVersions"`
 	}
 
-	if err := b.info.CallContext(ctx, &v, "info.getNodeVersion"); err != nil {
-		return nil, err
-	}
-
-	if avm, ok := v.VMVersions["avm"]; ok {
-		ci.Version = avm
+	if err := b.info.CallContext(ctx, &v, "info.getNodeVersion"); err == nil {
+		if avm, ok := v.VMVersions["avm"]; ok {
+			ci.Version = avm
+		}
 	}
 
 	return ci, nil
-}
-
-// EthereumTypeEstimateGas returns estimation of gas consumption for given transaction parameters
-func (b *AvalancheRPC) EthereumTypeEstimateGas(params map[string]interface{}) (uint64, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), b.Timeout)
-	defer cancel()
-	msg := interfaces.CallMsg{}
-	if s, ok := eth.GetStringFromMap("from", params); ok && len(s) > 0 {
-		msg.From = common.HexToAddress(s)
-	}
-	if s, ok := eth.GetStringFromMap("to", params); ok && len(s) > 0 {
-		a := common.HexToAddress(s)
-		msg.To = &a
-	}
-	if s, ok := eth.GetStringFromMap("data", params); ok && len(s) > 0 {
-		msg.Data = common.FromHex(s)
-	}
-	if s, ok := eth.GetStringFromMap("value", params); ok && len(s) > 0 {
-		msg.Value, _ = hexutil.DecodeBig(s)
-	}
-	if s, ok := eth.GetStringFromMap("gas", params); ok && len(s) > 0 {
-		msg.Gas, _ = hexutil.DecodeUint64(s)
-	}
-	if s, ok := eth.GetStringFromMap("gasPrice", params); ok && len(s) > 0 {
-		msg.GasPrice, _ = hexutil.DecodeBig(s)
-	}
-	return b.Client.EstimateGas(ctx, msg)
 }
