@@ -14,10 +14,10 @@ const (
 	// MainNet is production network
 	MainNet eth.Network = 56
 
-	// bsc token type names
-	BEP20TokenType   bchain.TokenTypeName = "BEP20"
-	BEP721TokenType  bchain.TokenTypeName = "BEP721"
-	BEP1155TokenType bchain.TokenTypeName = "BEP1155"
+	// bsc token standard names
+	BEP20TokenStandard   bchain.TokenStandardName = "BEP20"
+	BEP721TokenStandard  bchain.TokenStandardName = "BEP721"
+	BEP1155TokenStandard bchain.TokenStandardName = "BEP1155"
 )
 
 // BNBSmartChainRPC is an interface to JSON-RPC bsc service.
@@ -32,12 +32,13 @@ func NewBNBSmartChainRPC(config json.RawMessage, pushHandler func(bchain.Notific
 		return nil, err
 	}
 
-	// overwrite EthereumTokenTypeMap with bsc specific token type names
-	bchain.EthereumTokenTypeMap = []bchain.TokenTypeName{BEP20TokenType, BEP721TokenType, BEP1155TokenType}
+	// overwrite EthereumTokenStandardMap with bsc specific token standard names
+	bchain.EthereumTokenStandardMap = []bchain.TokenStandardName{BEP20TokenStandard, BEP721TokenStandard, BEP1155TokenStandard}
 
 	s := &BNBSmartChainRPC{
 		EthereumRPC: c.(*eth.EthereumRPC),
 	}
+	s.Parser.EnsSuffix = ".bnb"
 
 	return s, nil
 }
@@ -74,6 +75,8 @@ func (b *BNBSmartChainRPC) Initialize() error {
 	default:
 		return errors.Errorf("Unknown network id %v", id)
 	}
+
+	b.InitAlternativeProviders()
 
 	glog.Info("rpc: block chain ", b.Network)
 
