@@ -3,11 +3,11 @@ package ecash
 import (
 	"fmt"
 
-	"github.com/pirk/ecashutil"
 	"github.com/martinboehm/btcutil"
 	"github.com/martinboehm/btcutil/chaincfg"
 	"github.com/martinboehm/btcutil/txscript"
 	"github.com/pirk/ecashaddr-converter/address"
+	"github.com/pirk/ecashutil"
 	"github.com/trezor/blockbook/bchain"
 	"github.com/trezor/blockbook/bchain/coins/btc"
 )
