@@ -91,16 +91,19 @@ func init() {
 		},
 		CoinSpecificData: bchain.EthereumSpecificData{
 			Tx: &bchain.RpcTransaction{
-				AccountNonce:     "0xb26c",
-				GasPrice:         "0x430e23400",
-				GasLimit:         "0x5208",
-				To:               "0x555Ee11FBDDc0E49A9bAB358A8941AD95fFDB48f",
-				Value:            "0x1bc0159d530e6000",
-				Payload:          "0x",
-				Hash:             "0xcd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b",
-				BlockNumber:      "0x41eee8",
-				From:             "0x3E3a3D69dc66bA10737F531ed088954a9EC89d97",
-				TransactionIndex: "0xa",
+				AccountNonce:         "0xb26c",
+				GasPrice:             "0x430e23400",
+				MaxPriorityFeePerGas: "0x430e23401",
+				MaxFeePerGas:         "0x430e23402",
+				BaseFeePerGas:        "0x430e23403",
+				GasLimit:             "0x5208",
+				To:                   "0x555Ee11FBDDc0E49A9bAB358A8941AD95fFDB48f",
+				Value:                "0x1bc0159d530e6000",
+				Payload:              "0x",
+				Hash:                 "0xcd647151552b5132b2aef7c9be00dc6f73afc5901dde157aab131335baaa853b",
+				BlockNumber:          "0x41eee8",
+				From:                 "0x3E3a3D69dc66bA10737F531ed088954a9EC89d97",
+				TransactionIndex:     "0xa",
 			},
 			Receipt: &bchain.RpcReceipt{
 				GasUsed: "0x5208",
@@ -129,16 +132,19 @@ func init() {
 		},
 		CoinSpecificData: bchain.EthereumSpecificData{
 			Tx: &bchain.RpcTransaction{
-				AccountNonce:     "0xd0",
-				GasPrice:         "0x9502f9000",
-				GasLimit:         "0x130d5",
-				To:               "0x4af4114F73d1c1C903aC9E0361b379D1291808A2",
-				Value:            "0x0",
-				Payload:          "0xa9059cbb000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f00000000000000000000000000000000000000000000021e19e0c9bab2400000",
-				Hash:             "0xa9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b101",
-				BlockNumber:      "0x41eee8",
-				From:             "0x20cD153de35D469BA46127A0C8F18626b59a256A",
-				TransactionIndex: "0x0"},
+				AccountNonce:         "0xd0",
+				GasPrice:             "0x9502f9000",
+				MaxPriorityFeePerGas: "0x9502f9001",
+				MaxFeePerGas:         "0x9502f9002",
+				BaseFeePerGas:        "0x9502f9003",
+				GasLimit:             "0x130d5",
+				To:                   "0x4af4114F73d1c1C903aC9E0361b379D1291808A2",
+				Value:                "0x0",
+				Payload:              "0xa9059cbb000000000000000000000000555ee11fbddc0e49a9bab358a8941ad95ffdb48f00000000000000000000000000000000000000000000021e19e0c9bab2400000",
+				Hash:                 "0xa9cd088aba2131000da6f38a33c20169baee476218deea6b78720700b895b101",
+				BlockNumber:          "0x41eee8",
+				From:                 "0x20cD153de35D469BA46127A0C8F18626b59a256A",
+				TransactionIndex:     "0x0"},
 			Receipt: &bchain.RpcReceipt{
 				GasUsed: "0xcb39",
 				Status:  "0x1",
