@@ -14,13 +14,13 @@ import (
 )
 
 const (
-	OpZeroCoinMint      = 0xc1
-	OpZeroCoinSpend     = 0xc2
-	OpSigmaMint         = 0xc3
-	OpSigmaSpend        = 0xc4
-	OpLelantusMint      = 0xc5
-	OpLelantusJMint     = 0xc6
-	OpLelantusJoinSplit = 0xc7
+	OpZeroCoinMint             = 0xc1
+	OpZeroCoinSpend            = 0xc2
+	OpSigmaMint                = 0xc3
+	OpSigmaSpend               = 0xc4
+	OpLelantusMint             = 0xc5
+	OpLelantusJMint            = 0xc6
+	OpLelantusJoinSplit        = 0xc7
 	OpLelantusJoinSplitPayload = 0xc9
 
 	MainnetMagic wire.BitcoinNet = 0xe3d9fef1
@@ -194,7 +194,6 @@ func (p *FiroParser) ParseBlock(b []byte) (*bchain.Block, error) {
 					break
 				}
 			}
-			
 
 			if !isAllZero {
 				// hash data
@@ -344,7 +343,7 @@ type MTPHashDataRoot struct {
 }
 
 type MTPHashData struct {
-	BlockMTP    [128][128]uint64
+	BlockMTP [128][128]uint64
 }
 
 type MTPBlockHeader struct {
