@@ -7,6 +7,7 @@ import (
 	"os"
 	"sort"
 	"sync"
+	"time"
 
 	vlq "github.com/bsm/go-vlq"
 	"github.com/golang/glog"
@@ -108,7 +109,7 @@ func (s *MultiTokenValues) upsert(m bchain.MultiTokenValue, index int32, aggrega
 
 // AddrContract is Contract address with number of transactions done by given address
 type AddrContract struct {
-	Type             bchain.TokenType
+	Standard         bchain.TokenStandard
 	Contract         bchain.AddressDescriptor
 	Txs              uint
 	Value            big.Int          // single value of ERC20
@@ -124,8 +125,8 @@ type AddrContracts struct {
 	Contracts      []AddrContract
 }
 
-// packAddrContract packs AddrContracts into a byte buffer
-func packAddrContracts(acs *AddrContracts) []byte {
+// packAddrContracts packs AddrContracts into a byte buffer
+func packAddrContractsV6(acs *AddrContracts) []byte {
 	buf := make([]byte, 0, 128)
 	varBuf := make([]byte, maxPackedBigintBytes)
 	l := packVaruint(acs.TotalTxs, varBuf)
@@ -136,12 +137,12 @@ func packAddrContracts(acs *AddrContracts) []byte {
 	buf = append(buf, varBuf[:l]...)
 	for _, ac := range acs.Contracts {
 		buf = append(buf, ac.Contract...)
-		l = packVaruint(uint(ac.Type)+ac.Txs<<2, varBuf)
+		l = packVaruint(uint(ac.Standard)+ac.Txs<<2, varBuf)
 		buf = append(buf, varBuf[:l]...)
-		if ac.Type == bchain.FungibleToken {
+		if ac.Standard == bchain.FungibleToken {
 			l = packBigint(&ac.Value, varBuf)
 			buf = append(buf, varBuf[:l]...)
-		} else if ac.Type == bchain.NonFungibleToken {
+		} else if ac.Standard == bchain.NonFungibleToken {
 			l = packVaruint(uint(len(ac.Ids)), varBuf)
 			buf = append(buf, varBuf[:l]...)
 			for i := range ac.Ids {
@@ -162,14 +163,54 @@ func packAddrContracts(acs *AddrContracts) []byte {
 	return buf
 }
 
-func unpackAddrContracts(buf []byte, addrDesc bchain.AddressDescriptor) (*AddrContracts, error) {
+// packAddrContracts packs AddrContracts into a byte buffer
+func packAddrContracts(acs *AddrContracts) []byte {
+	buf := make([]byte, 0, 8+len(acs.Contracts)*(eth.EthereumTypeAddressDescriptorLen+12))
+	varBuf := make([]byte, maxPackedBigintBytes)
+	l := packVaruint(acs.TotalTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(acs.NonContractTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(acs.InternalTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(uint(len(acs.Contracts)), varBuf)
+	buf = append(buf, varBuf[:l]...)
+	for _, ac := range acs.Contracts {
+		buf = append(buf, ac.Contract...)
+		l = packVaruint(uint(ac.Standard)+ac.Txs<<2, varBuf)
+		buf = append(buf, varBuf[:l]...)
+		if ac.Standard == bchain.FungibleToken {
+			l = packBigint(&ac.Value, varBuf)
+			buf = append(buf, varBuf[:l]...)
+		} else if ac.Standard == bchain.NonFungibleToken {
+			l = packVaruint(uint(len(ac.Ids)), varBuf)
+			buf = append(buf, varBuf[:l]...)
+			for i := range ac.Ids {
+				l = packBigint(&ac.Ids[i], varBuf)
+				buf = append(buf, varBuf[:l]...)
+			}
+		} else { // bchain.ERC1155
+			l = packVaruint(uint(len(ac.MultiTokenValues)), varBuf)
+			buf = append(buf, varBuf[:l]...)
+			for i := range ac.MultiTokenValues {
+				l = packBigint(&ac.MultiTokenValues[i].Id, varBuf)
+				buf = append(buf, varBuf[:l]...)
+				l = packBigint(&ac.MultiTokenValues[i].Value, varBuf)
+				buf = append(buf, varBuf[:l]...)
+			}
+		}
+	}
+	return buf
+}
+
+func unpackAddrContractsV6(buf []byte, addrDesc bchain.AddressDescriptor) (acs *AddrContracts, err error) {
 	tt, l := unpackVaruint(buf)
 	buf = buf[l:]
 	nct, l := unpackVaruint(buf)
 	buf = buf[l:]
 	ict, l := unpackVaruint(buf)
 	buf = buf[l:]
-	c := make([]AddrContract, 0, 4)
+	c := make([]AddrContract, 0, len(buf)/30+4)
 	for len(buf) > 0 {
 		if len(buf) < eth.EthereumTypeAddressDescriptorLen {
 			return nil, errors.New("Invalid data stored in cfAddressContracts for AddrDesc " + addrDesc.String())
@@ -177,21 +218,81 @@ func unpackAddrContracts(buf []byte, addrDesc bchain.AddressDescriptor) (*AddrCo
 		contract := append(bchain.AddressDescriptor(nil), buf[:eth.EthereumTypeAddressDescriptorLen]...)
 		txs, l := unpackVaruint(buf[eth.EthereumTypeAddressDescriptorLen:])
 		buf = buf[eth.EthereumTypeAddressDescriptorLen+l:]
-		ttt := bchain.TokenType(txs & 3)
+		standard := bchain.TokenStandard(txs & 3)
 		txs >>= 2
 		ac := AddrContract{
-			Type:     ttt,
+			Standard: standard,
 			Contract: contract,
 			Txs:      txs,
 		}
-		if ttt == bchain.FungibleToken {
+		if standard == bchain.FungibleToken {
 			b, ll := unpackBigint(buf)
 			buf = buf[ll:]
 			ac.Value = b
 		} else {
 			len, ll := unpackVaruint(buf)
 			buf = buf[ll:]
-			if ttt == bchain.NonFungibleToken {
+			if standard == bchain.NonFungibleToken {
+				ac.Ids = make(Ids, len)
+				for i := uint(0); i < len; i++ {
+					b, ll := unpackBigint(buf)
+					buf = buf[ll:]
+					ac.Ids[i] = b
+				}
+			} else {
+				ac.MultiTokenValues = make(MultiTokenValues, len)
+				for i := uint(0); i < len; i++ {
+					b, ll := unpackBigint(buf)
+					buf = buf[ll:]
+					ac.MultiTokenValues[i].Id = b
+					b, ll = unpackBigint(buf)
+					buf = buf[ll:]
+					ac.MultiTokenValues[i].Value = b
+				}
+			}
+		}
+		c = append(c, ac)
+	}
+	return &AddrContracts{
+		TotalTxs:       tt,
+		NonContractTxs: nct,
+		InternalTxs:    ict,
+		Contracts:      c,
+	}, nil
+}
+
+func unpackAddrContracts(buf []byte, addrDesc bchain.AddressDescriptor) (acs *AddrContracts, err error) {
+	tt, l := unpackVaruint(buf)
+	buf = buf[l:]
+	nct, l := unpackVaruint(buf)
+	buf = buf[l:]
+	ict, l := unpackVaruint(buf)
+	buf = buf[l:]
+	cl, l := unpackVaruint(buf)
+	buf = buf[l:]
+	c := make([]AddrContract, 0, cl)
+	for len(buf) > 0 {
+		if len(buf) < eth.EthereumTypeAddressDescriptorLen {
+			return nil, errors.New("Invalid data stored in cfAddressContracts for AddrDesc " + addrDesc.String())
+		}
+		contract := append(bchain.AddressDescriptor(nil), buf[:eth.EthereumTypeAddressDescriptorLen]...)
+		txs, l := unpackVaruint(buf[eth.EthereumTypeAddressDescriptorLen:])
+		buf = buf[eth.EthereumTypeAddressDescriptorLen+l:]
+		standard := bchain.TokenStandard(txs & 3)
+		txs >>= 2
+		ac := AddrContract{
+			Standard: standard,
+			Contract: contract,
+			Txs:      txs,
+		}
+		if standard == bchain.FungibleToken {
+			b, ll := unpackBigint(buf)
+			buf = buf[ll:]
+			ac.Value = b
+		} else {
+			len, ll := unpackVaruint(buf)
+			buf = buf[ll:]
+			if standard == bchain.NonFungibleToken {
 				ac.Ids = make(Ids, len)
 				for i := uint(0); i < len; i++ {
 					b, ll := unpackBigint(buf)
@@ -247,7 +348,7 @@ func (d *RocksDB) GetAddrDescContracts(addrDesc bchain.AddressDescriptor) (*Addr
 	return unpackAddrContracts(buf, addrDesc)
 }
 
-func findContractInAddressContracts(contract bchain.AddressDescriptor, contracts []AddrContract) (int, bool) {
+func findContractInAddressContracts(contract bchain.AddressDescriptor, contracts []unpackedAddrContract) (int, bool) {
 	for i := range contracts {
 		if bytes.Equal(contract, contracts[i].Contract) {
 			return i, true
@@ -298,7 +399,7 @@ func addToAddressesMapEthereumType(addresses addressesMap, strAddrDesc string, b
 	return false
 }
 
-func addToContract(c *AddrContract, contractIndex int, index int32, contract bchain.AddressDescriptor, transfer *bchain.TokenTransfer, addTxCount bool) int32 {
+func addToContract(c *unpackedAddrContract, contractIndex int, index int32, contract bchain.AddressDescriptor, transfer *bchain.TokenTransfer, addTxCount bool) int32 {
 	var aggregate AggregateFn
 	// index 0 is for ETH transfers, index 1 (InternalTxIndexOffset) is for internal transfers, contract indexes start with 2 (ContractIndexOffset)
 	if index < 0 {
@@ -316,9 +417,9 @@ func addToContract(c *AddrContract, contractIndex int, index int32, contract bch
 			s.Add(s, v)
 		}
 	}
-	if transfer.Type == bchain.FungibleToken {
-		aggregate(&c.Value, &transfer.Value)
-	} else if transfer.Type == bchain.NonFungibleToken {
+	if transfer.Standard == bchain.FungibleToken {
+		aggregate(c.Value.get(), &transfer.Value)
+	} else if transfer.Standard == bchain.NonFungibleToken {
 		if index < 0 {
 			c.Ids.remove(transfer.Value)
 		} else {
@@ -335,17 +436,17 @@ func addToContract(c *AddrContract, contractIndex int, index int32, contract bch
 	return index
 }
 
-func (d *RocksDB) addToAddressesAndContractsEthereumType(addrDesc bchain.AddressDescriptor, btxID []byte, index int32, contract bchain.AddressDescriptor, transfer *bchain.TokenTransfer, addTxCount bool, addresses addressesMap, addressContracts map[string]*AddrContracts) error {
+func (d *RocksDB) addToAddressesAndContractsEthereumType(addrDesc bchain.AddressDescriptor, btxID []byte, index int32, contract bchain.AddressDescriptor, transfer *bchain.TokenTransfer, addTxCount bool, addresses addressesMap, addressContracts map[string]*unpackedAddrContracts) error {
 	var err error
 	strAddrDesc := string(addrDesc)
 	ac, e := addressContracts[strAddrDesc]
 	if !e {
-		ac, err = d.GetAddrDescContracts(addrDesc)
+		ac, err = d.getUnpackedAddrDescContracts(addrDesc)
 		if err != nil {
 			return err
 		}
 		if ac == nil {
-			ac = &AddrContracts{}
+			ac = &unpackedAddrContracts{}
 		}
 		addressContracts[strAddrDesc] = ac
 		d.cbs.balancesMiss++
@@ -367,9 +468,9 @@ func (d *RocksDB) addToAddressesAndContractsEthereumType(addrDesc bchain.Address
 			contractIndex, found := findContractInAddressContracts(contract, ac.Contracts)
 			if !found {
 				contractIndex = len(ac.Contracts)
-				ac.Contracts = append(ac.Contracts, AddrContract{
+				ac.Contracts = append(ac.Contracts, unpackedAddrContract{
 					Contract: contract,
-					Type:     transfer.Type,
+					Standard: transfer.Standard,
 				})
 			}
 			c := &ac.Contracts[contractIndex]
@@ -391,7 +492,7 @@ func (d *RocksDB) addToAddressesAndContractsEthereumType(addrDesc bchain.Address
 
 type ethBlockTxContract struct {
 	from, to, contract bchain.AddressDescriptor
-	transferType       bchain.TokenType
+	transferStandard   bchain.TokenStandard
 	value              big.Int
 	idValues           []bchain.MultiTokenValue
 }
@@ -416,7 +517,7 @@ type ethBlockTx struct {
 	internalData *ethInternalData
 }
 
-func (d *RocksDB) processBaseTxData(blockTx *ethBlockTx, tx *bchain.Tx, addresses addressesMap, addressContracts map[string]*AddrContracts) error {
+func (d *RocksDB) processBaseTxData(blockTx *ethBlockTx, tx *bchain.Tx, addresses addressesMap, addressContracts map[string]*unpackedAddrContracts) error {
 	var from, to bchain.AddressDescriptor
 	var err error
 	// there is only one output address in EthereumType transaction, store it in format txid 0
@@ -467,7 +568,7 @@ func (d *RocksDB) setAddressTxIndexesToAddressMap(addrDesc bchain.AddressDescrip
 }
 
 // existingBlock signals that internal data are reconnected to already indexed block after they failed during standard sync
-func (d *RocksDB) processInternalData(blockTx *ethBlockTx, tx *bchain.Tx, id *bchain.EthereumInternalData, addresses addressesMap, addressContracts map[string]*AddrContracts, existingBlock bool) error {
+func (d *RocksDB) processInternalData(blockTx *ethBlockTx, tx *bchain.Tx, id *bchain.EthereumInternalData, addresses addressesMap, addressContracts map[string]*unpackedAddrContracts, existingBlock bool) error {
 	blockTx.internalData = &ethInternalData{
 		internalType: id.Type,
 		errorMsg:     id.Error,
@@ -539,7 +640,7 @@ func (d *RocksDB) processInternalData(blockTx *ethBlockTx, tx *bchain.Tx, id *bc
 	return nil
 }
 
-func (d *RocksDB) processContractTransfers(blockTx *ethBlockTx, tx *bchain.Tx, addresses addressesMap, addressContracts map[string]*AddrContracts) error {
+func (d *RocksDB) processContractTransfers(blockTx *ethBlockTx, tx *bchain.Tx, addresses addressesMap, addressContracts map[string]*unpackedAddrContracts) error {
 	tokenTransfers, err := d.chainParser.EthereumTypeGetTokenTransfersFromTx(tx)
 	if err != nil {
 		glog.Warningf("rocksdb: processContractTransfers %v, tx %v", err, tx.Txid)
@@ -566,7 +667,7 @@ func (d *RocksDB) processContractTransfers(blockTx *ethBlockTx, tx *bchain.Tx, a
 			return err
 		}
 		bc := &blockTx.contracts[i]
-		bc.transferType = t.Type
+		bc.transferStandard = t.Standard
 		bc.from = from
 		bc.to = to
 		bc.contract = contract
@@ -576,7 +677,7 @@ func (d *RocksDB) processContractTransfers(blockTx *ethBlockTx, tx *bchain.Tx, a
 	return nil
 }
 
-func (d *RocksDB) processAddressesEthereumType(block *bchain.Block, addresses addressesMap, addressContracts map[string]*AddrContracts) ([]ethBlockTx, error) {
+func (d *RocksDB) processAddressesEthereumType(block *bchain.Block, addresses addressesMap, addressContracts map[string]*unpackedAddrContracts) ([]ethBlockTx, error) {
 	blockTxs := make([]ethBlockTx, len(block.Txs))
 	for txi := range block.Txs {
 		tx := &block.Txs[txi]
@@ -606,6 +707,9 @@ func (d *RocksDB) processAddressesEthereumType(block *bchain.Block, addresses ad
 
 // ReconnectInternalDataToBlockEthereumType adds missing internal data to the block and stores them in db
 func (d *RocksDB) ReconnectInternalDataToBlockEthereumType(block *bchain.Block) error {
+	d.connectBlockMux.Lock()
+	defer d.connectBlockMux.Unlock()
+
 	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
 	if d.chainParser.GetChainType() != bchain.ChainEthereumType {
@@ -613,7 +717,7 @@ func (d *RocksDB) ReconnectInternalDataToBlockEthereumType(block *bchain.Block) 
 	}
 
 	addresses := make(addressesMap)
-	addressContracts := make(map[string]*AddrContracts)
+	addressContracts := make(map[string]*unpackedAddrContracts)
 
 	// process internal data
 	blockTxs := make([]ethBlockTx, len(block.Txs))
@@ -634,7 +738,7 @@ func (d *RocksDB) ReconnectInternalDataToBlockEthereumType(block *bchain.Block) 
 		}
 	}
 
-	if err := d.storeAddressContracts(wb, addressContracts); err != nil {
+	if err := d.storeUnpackedAddressContracts(wb, addressContracts); err != nil {
 		return err
 	}
 	if err := d.storeInternalDataEthereumType(wb, blockTxs); err != nil {
@@ -847,7 +951,7 @@ var cachedContractsMux sync.Mutex
 func packContractInfo(contractInfo *bchain.ContractInfo) []byte {
 	buf := packString(contractInfo.Name)
 	buf = append(buf, packString(contractInfo.Symbol)...)
-	buf = append(buf, packString(string(contractInfo.Type))...)
+	buf = append(buf, packString(string(contractInfo.Standard))...)
 	varBuf := make([]byte, vlq.MaxLen64)
 	l := packVaruint(uint(contractInfo.Decimals), varBuf)
 	buf = append(buf, varBuf[:l]...)
@@ -868,7 +972,8 @@ func unpackContractInfo(buf []byte) (*bchain.ContractInfo, error) {
 	contractInfo.Symbol, l = unpackString(buf)
 	buf = buf[l:]
 	s, l = unpackString(buf)
-	contractInfo.Type = bchain.TokenTypeName(s)
+	contractInfo.Standard = bchain.TokenStandardName(s)
+	contractInfo.Type = bchain.TokenStandardName(s)
 	buf = buf[l:]
 	ui, l = unpackVaruint(buf)
 	contractInfo.Decimals = int(ui)
@@ -889,9 +994,9 @@ func (d *RocksDB) GetContractInfoForAddress(address string) (*bchain.ContractInf
 	return d.GetContractInfo(contract, "")
 }
 
-// GetContractInfo gets contract from cache or DB and possibly updates the type from typeFromContext
-// it is hard to guess the type of the contract using API, it is easier to set it the first time the contract is processed in a tx
-func (d *RocksDB) GetContractInfo(contract bchain.AddressDescriptor, typeFromContext bchain.TokenTypeName) (*bchain.ContractInfo, error) {
+// GetContractInfo gets contract from cache or DB and possibly updates the standard from standardFromContext
+// it is hard to guess the standard of the contract using API, it is easier to set it the first time the contract is processed in a tx
+func (d *RocksDB) GetContractInfo(contract bchain.AddressDescriptor, standardFromContext bchain.TokenStandardName) (*bchain.ContractInfo, error) {
 	cacheKey := string(contract)
 	cachedContractsMux.Lock()
 	contractInfo, found := cachedContracts[cacheKey]
@@ -911,9 +1016,10 @@ func (d *RocksDB) GetContractInfo(contract bchain.AddressDescriptor, typeFromCon
 		if len(addresses) > 0 {
 			contractInfo.Contract = addresses[0]
 		}
-		// if the type is specified and stored contractInfo has unknown type, set and store it
-		if typeFromContext != bchain.UnknownTokenType && contractInfo.Type == bchain.UnknownTokenType {
-			contractInfo.Type = typeFromContext
+		// if the standard is specified and stored contractInfo has unknown standard, set and store it
+		if standardFromContext != bchain.UnknownTokenStandard && contractInfo.Standard == bchain.UnknownTokenStandard {
+			contractInfo.Standard = standardFromContext
+			contractInfo.Type = standardFromContext
 			err = d.db.PutCF(d.wo, d.cfh[cfContracts], contract, packContractInfo(contractInfo))
 			if err != nil {
 				return nil, err
@@ -978,9 +1084,9 @@ func packBlockTx(buf []byte, blockTx *ethBlockTx) []byte {
 		buf = appendAddress(buf, c.from)
 		buf = appendAddress(buf, c.to)
 		buf = appendAddress(buf, c.contract)
-		l = packVaruint(uint(c.transferType), varBuf)
+		l = packVaruint(uint(c.transferStandard), varBuf)
 		buf = append(buf, varBuf[:l]...)
-		if c.transferType == bchain.MultiToken {
+		if c.transferStandard == bchain.MultiToken {
 			l = packVaruint(uint(len(c.idValues)), varBuf)
 			buf = append(buf, varBuf[:l]...)
 			for i := range c.idValues {
@@ -1142,9 +1248,9 @@ func unpackBlockTx(buf []byte, pos int) (*ethBlockTx, int, error) {
 			return nil, 0, err
 		}
 		cc, l = unpackVaruint(buf[pos:])
-		c.transferType = bchain.TokenType(cc)
+		c.transferStandard = bchain.TokenStandard(cc)
 		pos += l
-		if c.transferType == bchain.MultiToken {
+		if c.transferStandard == bchain.MultiToken {
 			cc, l = unpackVaruint(buf[pos:])
 			pos += l
 			c.idValues = make([]bchain.MultiTokenValue, cc)
@@ -1191,7 +1297,7 @@ func (d *RocksDB) getBlockTxsEthereumType(height uint32) ([]ethBlockTx, error) {
 	return bt, nil
 }
 
-func (d *RocksDB) disconnectAddress(btxID []byte, internal bool, addrDesc bchain.AddressDescriptor, btxContract *ethBlockTxContract, addresses map[string]map[string]struct{}, contracts map[string]*AddrContracts) error {
+func (d *RocksDB) disconnectAddress(btxID []byte, internal bool, addrDesc bchain.AddressDescriptor, btxContract *ethBlockTxContract, addresses map[string]map[string]struct{}, contracts map[string]*unpackedAddrContracts) error {
 	var err error
 	// do not process empty address
 	if len(addrDesc) == 0 {
@@ -1213,7 +1319,7 @@ func (d *RocksDB) disconnectAddress(btxID []byte, internal bool, addrDesc bchain
 	}
 	addrContracts, fc := contracts[s]
 	if !fc {
-		addrContracts, err = d.GetAddrDescContracts(addrDesc)
+		addrContracts, err = d.getUnpackedAddrDescContracts(addrDesc)
 		if err != nil {
 			return err
 		}
@@ -1257,7 +1363,7 @@ func (d *RocksDB) disconnectAddress(btxID []byte, internal bool, addrDesc bchain
 							index = transferTo
 						}
 						addToContract(addrContract, contractIndex, index, btxContract.contract, &bchain.TokenTransfer{
-							Type:             btxContract.transferType,
+							Standard:         btxContract.transferStandard,
 							Value:            btxContract.value,
 							MultiTokenValues: btxContract.idValues,
 						}, false)
@@ -1279,7 +1385,7 @@ func (d *RocksDB) disconnectAddress(btxID []byte, internal bool, addrDesc bchain
 	return nil
 }
 
-func (d *RocksDB) disconnectInternalData(btxID []byte, addresses map[string]map[string]struct{}, contracts map[string]*AddrContracts) error {
+func (d *RocksDB) disconnectInternalData(btxID []byte, addresses map[string]map[string]struct{}, contracts map[string]*unpackedAddrContracts) error {
 	internalData, err := d.getEthereumInternalData(btxID)
 	if err != nil {
 		return err
@@ -1318,7 +1424,7 @@ func (d *RocksDB) disconnectInternalData(btxID []byte, addresses map[string]map[
 	return nil
 }
 
-func (d *RocksDB) disconnectBlockTxsEthereumType(wb *grocksdb.WriteBatch, height uint32, blockTxs []ethBlockTx, contracts map[string]*AddrContracts) error {
+func (d *RocksDB) disconnectBlockTxsEthereumType(wb *grocksdb.WriteBatch, height uint32, blockTxs []ethBlockTx, contracts map[string]*unpackedAddrContracts) error {
 	glog.Info("Disconnecting block ", height, " containing ", len(blockTxs), " transactions")
 	addresses := make(map[string]map[string]struct{})
 	for i := range blockTxs {
@@ -1377,7 +1483,7 @@ func (d *RocksDB) DisconnectBlockRangeEthereumType(lower uint32, higher uint32) 
 	}
 	wb := grocksdb.NewWriteBatch()
 	defer wb.Destroy()
-	contracts := make(map[string]*AddrContracts)
+	contracts := make(map[string]*unpackedAddrContracts)
 	for height := higher; height >= lower; height-- {
 		if err := d.disconnectBlockTxsEthereumType(wb, height, blocks[height-lower], contracts); err != nil {
 			return err
@@ -1387,7 +1493,7 @@ func (d *RocksDB) DisconnectBlockRangeEthereumType(lower uint32, higher uint32) 
 		wb.DeleteCF(d.cfh[cfHeight], key)
 		wb.DeleteCF(d.cfh[cfBlockInternalDataErrors], key)
 	}
-	d.storeAddressContracts(wb, contracts)
+	d.storeUnpackedAddressContracts(wb, contracts)
 	err := d.WriteBatch(wb)
 	if err == nil {
 		d.is.RemoveLastBlockTimes(int(higher-lower) + 1)
@@ -1452,4 +1558,296 @@ func (d *RocksDB) SortAddressContracts(stop chan os.Signal) error {
 	}
 	glog.Infof("SortAddressContracts: finished - scanned %d rows, sorted %d ids and %d multi token value", rowCount, idsSortedCount, multiTokenValuesSortedCount)
 	return nil
+}
+
+type unpackedBigInt struct {
+	Slice []byte
+	Value *big.Int
+}
+type unpackedIds []unpackedBigInt
+
+type unpackedAddrContract struct {
+	Standard         bchain.TokenStandard
+	Contract         bchain.AddressDescriptor
+	Txs              uint
+	Value            unpackedBigInt           // single value of ERC20
+	Ids              unpackedIds              // multiple ERC721 tokens
+	MultiTokenValues unpackedMultiTokenValues // multiple ERC1155 tokens
+}
+
+func (b *unpackedBigInt) get() *big.Int {
+	if b.Value == nil {
+		if len(b.Slice) == 0 {
+			b.Value = big.NewInt(0)
+		} else {
+			bi, _ := unpackBigint(b.Slice)
+			b.Value = &bi
+		}
+	}
+	return b.Value
+}
+
+type unpackedAddrContracts struct {
+	Packed         []byte
+	TotalTxs       uint
+	NonContractTxs uint
+	InternalTxs    uint
+	Contracts      []unpackedAddrContract
+}
+
+func (s *unpackedIds) search(id big.Int) int {
+	// attempt to find id using a binary search
+	return sort.Search(len(*s), func(i int) bool {
+		return (*s)[i].get().CmpAbs(&id) >= 0
+	})
+}
+
+// insert id in ascending order
+func (s *unpackedIds) insert(id big.Int) {
+	i := s.search(id)
+	if i == len(*s) {
+		*s = append(*s, unpackedBigInt{Value: &id})
+	} else {
+		*s = append((*s)[:i+1], (*s)[i:]...)
+		(*s)[i] = unpackedBigInt{Value: &id}
+	}
+}
+
+func (s *unpackedIds) remove(id big.Int) {
+	i := s.search(id)
+	// remove id if found
+	if i < len(*s) && (*s)[i].get().CmpAbs(&id) == 0 {
+		*s = append((*s)[:i], (*s)[i+1:]...)
+	}
+}
+
+type unpackedMultiTokenValue struct {
+	Id    unpackedBigInt
+	Value unpackedBigInt
+}
+
+type unpackedMultiTokenValues []unpackedMultiTokenValue
+
+// search for multi token value using a binary seach on id
+func (s *unpackedMultiTokenValues) search(m bchain.MultiTokenValue) int {
+	return sort.Search(len(*s), func(i int) bool {
+		return (*s)[i].Id.get().CmpAbs(&m.Id) >= 0
+	})
+}
+
+func (s *unpackedMultiTokenValues) upsert(m bchain.MultiTokenValue, index int32, aggregate AggregateFn) {
+	i := s.search(m)
+	if i < len(*s) && (*s)[i].Id.get().CmpAbs(&m.Id) == 0 {
+		aggregate((*s)[i].Value.get(), &m.Value)
+		// if transfer from, remove if the value is zero
+		if index < 0 && len((*s)[i].Value.get().Bits()) == 0 {
+			*s = append((*s)[:i], (*s)[i+1:]...)
+		}
+		return
+	}
+	if index >= 0 {
+		elem := unpackedMultiTokenValue{
+			Id:    unpackedBigInt{Value: &m.Id},
+			Value: unpackedBigInt{Value: new(big.Int).Set(&m.Value)},
+		}
+		if i == len(*s) {
+			*s = append(*s, elem)
+		} else {
+			*s = append((*s)[:i+1], (*s)[i:]...)
+			(*s)[i] = elem
+		}
+	}
+}
+
+// getUnpackedAddrDescContracts returns partially unpacked AddrContracts for given addrDesc
+func (d *RocksDB) getUnpackedAddrDescContracts(addrDesc bchain.AddressDescriptor) (*unpackedAddrContracts, error) {
+	d.addrContractsCacheMux.Lock()
+	rv, found := d.addrContractsCache[string(addrDesc)]
+	d.addrContractsCacheMux.Unlock()
+	if found && rv != nil {
+		return rv, nil
+	}
+	val, err := d.db.GetCF(d.ro, d.cfh[cfAddressContracts], addrDesc)
+	if err != nil {
+		return nil, err
+	}
+	defer val.Free()
+	buf := val.Data()
+	if len(buf) == 0 {
+		return nil, nil
+	}
+	rv, err = partiallyUnpackAddrContracts(buf)
+	if err == nil && rv != nil && len(buf) > addrContractsCacheMinSize {
+		d.addrContractsCacheMux.Lock()
+		d.addrContractsCache[string(addrDesc)] = rv
+		d.addrContractsCacheMux.Unlock()
+	}
+	return rv, err
+}
+
+// to speed up import of blocks, the unpacking of big ints is deferred to time when they are needed
+func partiallyUnpackAddrContracts(buf []byte) (acs *unpackedAddrContracts, err error) {
+	// make copy of the slice to avoid subsequent allocation of smaller slices
+	buf = append([]byte{}, buf...)
+	index := 0
+	tt, l := unpackVaruint(buf)
+	index += l
+	nct, l := unpackVaruint(buf[index:])
+	index += l
+	ict, l := unpackVaruint(buf[index:])
+	index += l
+	cl, l := unpackVaruint(buf[index:])
+	index += l
+	c := make([]unpackedAddrContract, 0, cl)
+	for index < len(buf) {
+		contract := buf[index : index+eth.EthereumTypeAddressDescriptorLen]
+		index += eth.EthereumTypeAddressDescriptorLen
+		txs, l := unpackVaruint(buf[index:])
+		index += l
+		standard := bchain.TokenStandard(txs & 3)
+		txs >>= 2
+		ac := unpackedAddrContract{
+			Standard: standard,
+			Contract: contract,
+			Txs:      txs,
+		}
+		if standard == bchain.FungibleToken {
+			l := packedBigintLen(buf[index:])
+			ac.Value = unpackedBigInt{Slice: buf[index : index+l]}
+			index += l
+		} else {
+			len, ll := unpackVaruint(buf[index:])
+			index += ll
+			if standard == bchain.NonFungibleToken {
+				ac.Ids = make(unpackedIds, len)
+				for i := uint(0); i < len; i++ {
+					ll := packedBigintLen(buf[index:])
+					ac.Ids[i] = unpackedBigInt{Slice: buf[index : index+ll]}
+					index += ll
+				}
+			} else {
+				ac.MultiTokenValues = make(unpackedMultiTokenValues, len)
+				for i := uint(0); i < len; i++ {
+					ll := packedBigintLen(buf[index:])
+					ac.MultiTokenValues[i].Id = unpackedBigInt{Slice: buf[index : index+ll]}
+					index += ll
+					ll = packedBigintLen(buf[index:])
+					ac.MultiTokenValues[i].Value = unpackedBigInt{Slice: buf[index : index+ll]}
+					index += ll
+				}
+			}
+		}
+		c = append(c, ac)
+	}
+	return &unpackedAddrContracts{
+		Packed:         buf,
+		TotalTxs:       tt,
+		NonContractTxs: nct,
+		InternalTxs:    ict,
+		Contracts:      c,
+	}, nil
+}
+
+// packUnpackedAddrContracts packs unpackedAddrContracts into a byte buffer
+func packUnpackedAddrContracts(acs *unpackedAddrContracts) []byte {
+	buf := make([]byte, 0, len(acs.Packed)+eth.EthereumTypeAddressDescriptorLen+12)
+	varBuf := make([]byte, maxPackedBigintBytes)
+	l := packVaruint(acs.TotalTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(acs.NonContractTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(acs.InternalTxs, varBuf)
+	buf = append(buf, varBuf[:l]...)
+	l = packVaruint(uint(len(acs.Contracts)), varBuf)
+	buf = append(buf, varBuf[:l]...)
+	for _, ac := range acs.Contracts {
+		buf = append(buf, ac.Contract...)
+		l = packVaruint(uint(ac.Standard)+ac.Txs<<2, varBuf)
+		buf = append(buf, varBuf[:l]...)
+		if ac.Standard == bchain.FungibleToken {
+			if ac.Value.Value != nil {
+				l = packBigint(ac.Value.Value, varBuf)
+				buf = append(buf, varBuf[:l]...)
+			} else {
+				buf = append(buf, ac.Value.Slice...)
+			}
+		} else if ac.Standard == bchain.NonFungibleToken {
+			l = packVaruint(uint(len(ac.Ids)), varBuf)
+			buf = append(buf, varBuf[:l]...)
+			for i := range ac.Ids {
+				if ac.Ids[i].Value != nil {
+					l = packBigint(ac.Ids[i].Value, varBuf)
+					buf = append(buf, varBuf[:l]...)
+				} else {
+					buf = append(buf, ac.Ids[i].Slice...)
+				}
+			}
+		} else { // bchain.ERC1155
+			l = packVaruint(uint(len(ac.MultiTokenValues)), varBuf)
+			buf = append(buf, varBuf[:l]...)
+			for i := range ac.MultiTokenValues {
+				if ac.MultiTokenValues[i].Id.Value != nil {
+					l = packBigint(ac.MultiTokenValues[i].Id.Value, varBuf)
+					buf = append(buf, varBuf[:l]...)
+				} else {
+					buf = append(buf, ac.MultiTokenValues[i].Id.Slice...)
+				}
+				if ac.MultiTokenValues[i].Value.Value != nil {
+					l = packBigint(ac.MultiTokenValues[i].Value.Value, varBuf)
+					buf = append(buf, varBuf[:l]...)
+				} else {
+					buf = append(buf, ac.MultiTokenValues[i].Value.Slice...)
+				}
+			}
+		}
+	}
+	return buf
+}
+
+func (d *RocksDB) storeUnpackedAddressContracts(wb *grocksdb.WriteBatch, acm map[string]*unpackedAddrContracts) error {
+	for addrDesc, acs := range acm {
+		// address with 0 contracts is removed from db - happens on disconnect
+		if acs == nil || (acs.NonContractTxs == 0 && acs.InternalTxs == 0 && len(acs.Contracts) == 0) {
+			wb.DeleteCF(d.cfh[cfAddressContracts], bchain.AddressDescriptor(addrDesc))
+		} else {
+			// do not store large address contracts found in cache
+			if _, found := d.addrContractsCache[addrDesc]; !found {
+				buf := packUnpackedAddrContracts(acs)
+				wb.PutCF(d.cfh[cfAddressContracts], bchain.AddressDescriptor(addrDesc), buf)
+			}
+		}
+	}
+	return nil
+}
+
+func (d *RocksDB) writeContractsCache() {
+	wb := grocksdb.NewWriteBatch()
+	defer wb.Destroy()
+	d.addrContractsCacheMux.Lock()
+	for addrDesc, acs := range d.addrContractsCache {
+		buf := packUnpackedAddrContracts(acs)
+		wb.PutCF(d.cfh[cfAddressContracts], bchain.AddressDescriptor(addrDesc), buf)
+	}
+	d.addrContractsCacheMux.Unlock()
+	if err := d.WriteBatch(wb); err != nil {
+		glog.Error("writeContractsCache: failed to store addrContractsCache: ", err)
+	}
+}
+
+func (d *RocksDB) storeAddrContractsCache() {
+	start := time.Now()
+	if len(d.addrContractsCache) > 0 {
+		d.writeContractsCache()
+	}
+	glog.Info("storeAddrContractsCache: store ", len(d.addrContractsCache), " entries in ", time.Since(start))
+}
+
+func (d *RocksDB) periodicStoreAddrContractsCache() {
+	period := time.Duration(5) * time.Minute
+	timer := time.NewTimer(period)
+	for {
+		<-timer.C
+		timer.Reset(period)
+		d.storeAddrContractsCache()
+	}
 }
