@@ -59,7 +59,7 @@ type marketChartPrices struct {
 }
 
 // NewCoinGeckoDownloader creates a coingecko structure that implements the RatesDownloaderInterface
-func NewCoinGeckoDownloader(db *db.RocksDB, coinShortcut string, url string, coin string, platformIdentifier string, platformVsCurrency string, allowedVsCurrencies string, timeFormat string, metrics *common.Metrics, throttleDown bool) RatesDownloaderInterface {
+func NewCoinGeckoDownloader(db *db.RocksDB, network string, url string, coin string, platformIdentifier string, platformVsCurrency string, allowedVsCurrencies string, timeFormat string, metrics *common.Metrics, throttleDown bool) RatesDownloaderInterface {
 	throttlingDelayMs := 0 // No delay by default
 	if throttleDown {
 		throttlingDelayMs = DefaultThrottleDelayMs
@@ -67,7 +67,7 @@ func NewCoinGeckoDownloader(db *db.RocksDB, coinShortcut string, url string, coi
 
 	allowedVsCurrenciesMap := getAllowedVsCurrenciesMap(allowedVsCurrencies)
 
-	apiKey := os.Getenv(strings.ToUpper(coinShortcut) + "_COINGECKO_API_KEY")
+	apiKey := os.Getenv(strings.ToUpper(network) + "_COINGECKO_API_KEY")
 	if apiKey == "" {
 		apiKey = os.Getenv("COINGECKO_API_KEY")
 	}
@@ -377,7 +377,7 @@ func (cg *Coingecko) HourlyTickers() (*[]common.CurrencyRatesTicker, error) {
 	return cg.getHighGranularityTickers("90")
 }
 
-// HourlyTickers returns the array of the exchange rates in five minutes granularity
+// FiveMinutesTickers returns the array of the exchange rates in five minutes granularity
 func (cg *Coingecko) FiveMinutesTickers() (*[]common.CurrencyRatesTicker, error) {
 	return cg.getHighGranularityTickers("1")
 }
