@@ -108,7 +108,7 @@ func NewFiatRates(db *db.RocksDB, config *common.Config, metrics *common.Metrics
 			// a small hack - in tests the callback is not used, therefore there is no delay slowing down the test
 			throttle = false
 		}
-		fr.downloader = NewCoinGeckoDownloader(db, db.GetInternalState().CoinShortcut, rdParams.URL, rdParams.Coin, rdParams.PlatformIdentifier, rdParams.PlatformVsCurrency, fr.allowedVsCurrencies, fr.timeFormat, metrics, throttle)
+		fr.downloader = NewCoinGeckoDownloader(db, db.GetInternalState().GetNetwork(), rdParams.URL, rdParams.Coin, rdParams.PlatformIdentifier, rdParams.PlatformVsCurrency, fr.allowedVsCurrencies, fr.timeFormat, metrics, throttle)
 		if is != nil {
 			is.HasFiatRates = true
 			is.HasTokenFiatRates = fr.downloadTokens
@@ -372,7 +372,7 @@ func (fr *FiatRates) tickersToMap(tickers *[]common.CurrencyRatesTicker, granula
 	return m, from, to
 }
 
-// setCurrentTicker sets hourly tickers
+// setHourlyTickers sets hourly tickers
 func (fr *FiatRates) setHourlyTickers(t *[]common.CurrencyRatesTicker) {
 	fr.db.FiatRatesStoreSpecialTickers(hourlyTickersKey, t)
 	fr.mux.Lock()
@@ -380,7 +380,7 @@ func (fr *FiatRates) setHourlyTickers(t *[]common.CurrencyRatesTicker) {
 	fr.hourlyTickers, fr.hourlyTickersFrom, fr.hourlyTickersTo = fr.tickersToMap(t, secondsInHour)
 }
 
-// setCurrentTicker sets hourly tickers
+// setFiveMinutesTickers sets five minutes tickers
 func (fr *FiatRates) setFiveMinutesTickers(t *[]common.CurrencyRatesTicker) {
 	fr.db.FiatRatesStoreSpecialTickers(fiveMinutesTickersKey, t)
 	fr.mux.Lock()
