@@ -3,6 +3,7 @@ package server
 import (
 	"encoding/json"
 	"fmt"
+	"html"
 	"html/template"
 	"math/big"
 	"net/http"
@@ -33,6 +34,66 @@ type htmlTemplates[TD any] struct {
 	newTemplateDataWithError func(error *api.APIError, r *http.Request) *TD
 	parseTemplates           func() []*template.Template
 	postHtmlTemplateHandler  func(data *TD, w http.ResponseWriter, r *http.Request)
+}
+
+func (s *htmlTemplates[TD]) jsonHandler(handler func(r *http.Request, apiVersion int) (interface{}, error), apiVersion int) func(w http.ResponseWriter, r *http.Request) {
+	type jsonError struct {
+		Text       string `json:"error"`
+		HTTPStatus int    `json:"-"`
+	}
+	handlerName := getFunctionName(handler)
+	return func(w http.ResponseWriter, r *http.Request) {
+		var data interface{}
+		var err error
+		defer func() {
+			if e := recover(); e != nil {
+				glog.Error(handlerName, " recovered from panic: ", e)
+				debug.PrintStack()
+				if s.debug {
+					data = jsonError{fmt.Sprint("Internal server error: recovered from panic ", e), http.StatusInternalServerError}
+				} else {
+					data = jsonError{"Internal server error", http.StatusInternalServerError}
+				}
+			}
+			w.Header().Set("Content-Type", "application/json; charset=utf-8")
+			if e, isError := data.(jsonError); isError {
+				w.WriteHeader(e.HTTPStatus)
+			}
+			err = json.NewEncoder(w).Encode(data)
+			if err != nil {
+				glog.Warning("json encode ", err)
+			}
+			if s.metrics != nil {
+				s.metrics.ExplorerPendingRequests.With((common.Labels{"method": handlerName})).Dec()
+			}
+		}()
+		if s.metrics != nil {
+			s.metrics.ExplorerPendingRequests.With((common.Labels{"method": handlerName})).Inc()
+		}
+		data, err = handler(r, apiVersion)
+		if err != nil || data == nil {
+			if apiErr, ok := err.(*api.APIError); ok {
+				if apiErr.Public {
+					data = jsonError{apiErr.Error(), http.StatusBadRequest}
+				} else {
+					data = jsonError{apiErr.Error(), http.StatusInternalServerError}
+				}
+			} else {
+				if err != nil {
+					glog.Error(handlerName, " error: ", err)
+				}
+				if s.debug {
+					if data != nil {
+						data = jsonError{fmt.Sprintf("Internal server error: %v, data %+v", err, data), http.StatusInternalServerError}
+					} else {
+						data = jsonError{fmt.Sprintf("Internal server error: %v", err), http.StatusInternalServerError}
+					}
+				} else {
+					data = jsonError{"Internal server error", http.StatusInternalServerError}
+				}
+			}
+		}
+	}
 }
 
 func (s *htmlTemplates[TD]) htmlTemplateHandler(handler func(w http.ResponseWriter, r *http.Request) (tpl, *TD, error)) func(w http.ResponseWriter, r *http.Request) {
@@ -214,7 +275,7 @@ func appendAmountSpan(rv *strings.Builder, class, amount, shortcut, txDate strin
 	}
 	if shortcut != "" {
 		rv.WriteString(" ")
-		rv.WriteString(shortcut)
+		rv.WriteString(html.EscapeString(shortcut))
 	}
 	rv.WriteString("</span>")
 }
@@ -257,7 +318,7 @@ func appendAmountSpanBitcoinType(rv *strings.Builder, class, amount, shortcut, t
 	rv.WriteString("</span>")
 	if shortcut != "" {
 		rv.WriteString(" ")
-		rv.WriteString(shortcut)
+		rv.WriteString(html.EscapeString(shortcut))
 	}
 	rv.WriteString("</span>")
 }
@@ -271,7 +332,7 @@ func appendAmountWrapperSpan(rv *strings.Builder, primary, symbol, classes strin
 	rv.WriteString(`" cc="`)
 	rv.WriteString(primary)
 	rv.WriteString(" ")
-	rv.WriteString(symbol)
+	rv.WriteString(html.EscapeString(symbol))
 	rv.WriteString(`">`)
 }
 
